Update
Forum page now properly loads in. The messages get retrieved in reversed order, so the newest messages get shown first instead of oldest first. The retrieved list of messages finally get reversed so they show up from oldest to newest message on the forum page, this means that the newest messages get shown on the bottom of the page instead of on the top.