Update
Fixed it so games can now be added. There was a mistake in the addGame function that it didn't check if the game already existed properly, and the return values were wrong too.

Added in a small Tester module. Use is for having a scheduled spam of creating new messages and forum posts